♻️ Simplify DD storage and reuse reset capacity
🤖 *AI text below* 🤖

Reuse a consolidated slab during reset(true). Replacing unchanged capacity
adds allocation churn and can terminate a memory-limited process needlessly.
Extend the allocator regression to cover repeated consolidation.

Use integer pool growth, standard bucket clearing, direct immortalization
of the sole stored constant, and automatic storage in the Grover test.
Refresh the decision record with the comparison against current main.

+++ include/mqt-core/dd/MemoryManager.hpp
@@ -44,7 +44,7 @@ public:
   static constexpr std::size_t INITIAL_ALLOCATION_SIZE = 2048U;
 
   /// Capacity multiplier when allocating the next chunk.
-  static constexpr double GROWTH_FACTOR = 2U;
+  static constexpr size_t GROWTH_FACTOR = 2U;
 
   /// Construct a new MemoryManager object for objects of type T.
   /// @param initialAllocationSize The initial number of entries to allocate
